add storybook-react; comes with a few example stories and assets
I had to delete the yarn.lock and reinstall the node modules to get this working, which is why there's so much changed in that file. I've run the site locally and can't see any issues. I also had a quick scan of the generated lock file and there seem to be no major version changes.